Move GetRecognizedMobileDevice() to pkg/lib/web
