+++ pkg/api/model/user_agent.go
@@ -57,30 +57,6 @@ func ParseUserAgent(ua string) (mUA UserAgent) {
 	return mUA
 }
 
-// The name is borrowed from https://github.com/browserslist/browserslist
-type RecognizedMobileDevice string
-
-const (
-	RecognizedMobileDeviceIOS           = "iOS"
-	RecognizedMobileDeviceChromeAndroid = "ChromeAndroid"
-	RecognizedMobileDeviceChrome        = "Chrome"
-	RecognizedMobileDeviceSamsung       = "Samsung"
-)
-
-func GetRecognizedMobileDevice(ua string) (string, bool) {
-	client := uaParser.Parse(ua)
-	if client.Os.Family == "iOS" {
-		return RecognizedMobileDeviceIOS, true
-	}
-	if client.Os.Family == "Android" && client.Device.Brand == "Samsung" {
-		return RecognizedMobileDeviceSamsung, true
-	}
-	if client.Os.Family == "Android" {
-		return RecognizedMobileDeviceChromeAndroid, true
-	}
-	if client.UserAgent.Family == "Chrome" {
-		return RecognizedMobileDeviceChrome, true
-	}
-
-	return "", false
+func ParseUserAgentRaw(ua string) *uaparser.Client {
+	return uaParser.Parse(ua)
 }
